+++ cpp/preamble/adapt.h
