+++ cpp/util/cli.h
