+++ cpp/util/iter.h
