+++ cpp/util/mem.h
