+++ cpp/util/mod.h
