--- harmonize/__init__.py
@@ -1,18 +1,0 @@
-# abs import
-import harmonize.python.config as config
-
-from harmonize.python.runtime import (
-    RuntimeSpec,
-)
-
-from harmonize.python.atomics   import (
-    array_atomic_add,
-    array_atomic_max,
-)
-
-from harmonize.python.printing  import print_formatted
-from harmonize.python.timing    import get_wall_clock
-
-from harmonize.python.array     import local_array
-
-
