+++ python/harmonize/templates.py
