+++ python/harmonize/timing.py
@@ -1,5 +1,5 @@
 import numba
-from harmonize.python.templates import clock_template
+from harmonize.templates import clock_template
 
 def generate_clock_code():
     return clock_template
